+++ src/wsjcpp_core.h
@@ -266,6 +266,7 @@ bool starts_with(const std::string &str, const std::string &start_str);
 bool ends_with(const std::string &str, const std::string &end_str);
 std::vector<std::string> directory_list(const std::string &dirpath);
 std::string generate_uuid();
+std::string parent_dirpath(const std::string &filepath);
 
 } // namespace wsjcpp