+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Fixed #40: Missing first character in name of child directory
 - Removed definition `WsjcppCore::listOfFiles` because missing implementation
 - Fixed #36: Renamed `WsjcppCore::createUuid` to `wsjcpp::generate_uuid()`
-
+- Fixed #33: Added `wsjcpp::parent_dirpath`
 
 ## [v0.2.5] - 2026-05-19 (2026 May 19)